requirements.txt
fastapi[standard]==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
google-genai==0.3.0
streamlit==1.39.0
scikit-learn==1.5.2
pandas==2.2.3
nltk==3.9.1
joblib==1.4.2
requests==2.32.3
SQLAlchemy==2.0.36
pymysql==1.1.1
apscheduler==3.11.1
cryptography
mysql-connector-python==9.5.0